+++ src/Pipes.php
@@ -1,160 +1,73 @@
 <?php
 namespace FutureProcess;
 
+use React\EventLoop\LoopInterface;
+use React\Stream\Stream;
+
 /**
  * @author Josh Di Fabio <joshdifabio@gmail.com>
  */
 class Pipes
 {
-    private static $modesByType = array(
-        'read' => array('r+', 'w', 'w+', 'a', 'a+', 'x', 'x+', 'c', 'c+'),
-        'write' => array('r', 'r+', 'w+', 'a+', 'x+', 'c+'),
-    );
-    
-    private $resources = array();
-    private $resourcesByType = array(
-        'read' => array(),
-        'write' => array(),
-    );
-    private $buffers = array(
-        'read' => array(),
-        'write' => array(),
-    );
-    
-    public function __construct(array $descriptorSpec)
+    private $eventLoop;
+    private $collectData;
+    private $pipes = array();
+    private $data = array();
+
+    public function __construct(LoopInterface $eventLoop, $collectData = true)
     {
-        $pipes = array_filter($descriptorSpec, function ($element) {
-            return isset($element[0]) && $element[0] === 'pipe';
-        });
-        
-        foreach (self::$modesByType as $type => $modes) {
-            $matchedPipes = array_filter($pipes, function ($pipe) use ($modes) {
-                return in_array($pipe[1], $modes);
-            });
-            
-            $this->buffers[$type] = array_fill_keys(array_keys($matchedPipes), '');
-        }
+        $this->eventLoop = $eventLoop;
+        $this->collectData = $collectData;
     }
-    
+
     public function setResources(array $resources)
     {
-        foreach (self::$modesByType as $type => $modes) {
-            $this->resourcesByType[$type] = array_intersect_key($resources, $this->buffers[$type]);
+        foreach ($resources as $descriptor => $resource) {
+            $this->initPipe($descriptor, $resource);
         }
-        
-        $this->resources = $resources;
     }
-    
-    public function getResource($descriptor)
+
+    public function getPipe($descriptor)
     {
-        if (!isset($this->resources[$descriptor])) {
+        if (!isset($this->pipes[$descriptor])) {
             throw new \RuntimeException('No pipe exists for the specified descriptor.');
         }
-        
-        return $this->resources[$descriptor];
+
+        return $this->pipes[$descriptor];
     }
-    
-    public function read($descriptor, $length = null)
+
+    public function getData($descriptor)
     {
-        if (!isset($this->buffers['read'][$descriptor])) {
+        if (!isset($this->pipes[$descriptor])) {
             throw new \RuntimeException('No pipe exists for the specified descriptor.');
         }
-        
-        if ($readResources = $this->resourcesByType['read']) {
-            $this->select($readResources, $writeResources);
-            $this->drainProcessOutputBuffers($readResources);
+
+        if (!isset($this->data[$descriptor])) {
+            throw new \RuntimeException('Data is not being collected for the specified descriptor.');
         }
-        
-        if (is_int($length)) {
-            $data = substr($this->buffers['read'][$descriptor], 0, $length);
-            $this->buffers['read'][$descriptor] = substr($this->buffers['read'][$descriptor], $length);
-        } else {
-            $data = $this->buffers['read'][$descriptor];
-            $this->buffers['read'][$descriptor] = '';
-        }
-        
-        return $data;
+
+        return $this->data[$descriptor];
     }
-    
-    public function write($descriptor, $data)
-    {
-        if (!isset($this->buffers['write'][$descriptor])) {
-            throw new \RuntimeException('No pipe exists for the specified descriptor.');
-        }
-        
-        $this->buffers['write'][$descriptor] .= $data;
-        
-        if ($writeResources = $this->resourcesByType['write']) {
-            $this->select($readResources, $writeResources);
-            $this->drainWriteBuffers($writeResources);
-        }
-    }
-    
-    public function readAndWrite()
-    {
-        $readResources = $this->resourcesByType['read'];
-        $writeResources = $this->resourcesByType['write'];
-        $this->select($readResources, $writeResources);
-        $this->drainWriteBuffers($writeResources);
-        $this->drainProcessOutputBuffers($readResources);
-    }
-    
+
     public function close()
     {
-        if ($readResources = $this->resourcesByType['read']) {
-            $this->select($readResources, $writeResources);
-            $this->drainProcessOutputBuffers($readResources);
-        }
-        
-        foreach ($this->resources as $descriptor => $resource) {
-            fclose($resource);
-            unset($this->resources[$descriptor]);
-            unset($this->resourcesByType['read'][$descriptor]);
-            unset($this->resourcesByType['write'][$descriptor]);
+        foreach ($this->pipes as $pipe) {
+            $pipe->close();
         }
     }
-    
-    private function drainProcessOutputBuffers(array $resources)
+
+    private function initPipe($descriptor, $resource)
     {
-        foreach ($resources as $descriptor => $resource) {
-            stream_set_blocking($resource, 0);
-            while (strlen($data = fread($resource, 8192))) {
-                $this->buffers['read'][$descriptor] .= $data;
-            }
+        $stream = new Stream($resource, $this->eventLoop);
+
+        if ($this->collectData) {
+            $dataArray = &$this->data;
+            $dataArray[$descriptor] = '';
+            $stream->on('data', function ($data) use (&$dataArray, $descriptor) {
+                $dataArray[$descriptor] .= $data;
+            });
         }
-    }
-    
-    private function drainWriteBuffers(array $resources)
-    {
-        foreach ($resources as $descriptor => $resource) {
-            stream_set_blocking($resource, 0);
-            $descriptor = array_search($resource, $this->resourcesByType);
-            while (strlen($this->buffers['write'][$descriptor])) {
-                $written = fwrite($resource, $this->buffers['write'][$descriptor], 2 << 18); // write 512k
-                if ($written > 0) {
-                    $this->buffers['write'][$descriptor] = (string)substr($this->buffers['write'][$descriptor], $written);
-                } else {
-                    break;
-                }
-            }
-        }
-    }
-    
-    private function select(&$readResources, &$writeResources)
-    {
-        if (false === stream_select($readResources, $writeResources, $except, 0)) {
-            throw new \RuntimeException('An error occurred when polling process pipes.');
-        }
-        
-        // prior PHP 5.4 the array passed to stream_select is modified and
-        // lose key association, we have to find back the key
-        
-        if (count($readResources)) {
-            $readResources = array_intersect($this->resources, $readResources);
-        }
-        
-        if (count($writeResources)) {
-            $writeResources = array_intersect($this->resources, $writeResources);
-        }
+
+        $this->pipes[$descriptor] = $stream;
     }
 }